ATO-980: Adds authenticated claim to JAR
This adds the authenticated claim to the JAR the stub sends to authorize
endpoint. This matches the updates made to this interface in orchestration[1]
and means any subsquent reliance on this field won't break the stub.

I've not modifed the session setup this stub does because I'm unsure if this
modifes the frontend session or the shared backend session.

[1]-https://github.com/govuk-one-login/authentication-api/pull/5310/files