AUT-3820: Parse rather than validate the request object
We either return a string representing an error or a domain object from the function that was previously just returning a true/false value. This gives us some benefits: we can return specific error messages based on which element of parsing has failed, and once we've validated/parsed, we can successfully just use the domain object.

We're doing a bit of parsing here of things that we haven't validated yet - that to come later