AUT-3834: Add continue button for ipv stub
This will initiate the callback logic to the frontend. As part of this we add a post endpoint for the stub which redirects back to the frontend's ipv-callback endpoint. We'll need to further refine the post to ensure that the request has a state that we then retrieve from the database, but for now we just do the simplest thing.