AUT-3681: Validate that the request contains a state field
This adds basic validation just that the claim is present (without checking any expected format, which we may want to do later. This has been added so that we can add state to the type that is returned by the parser function. This in turn means that we can store the state in the database, which we'll need to do alongside the auth code. This is a building block for eventually building out the token endpoint, which will need various things stored and exchange in the database.