AUT-3861: Add database for stub and store auth code with state
Adds the first database call required for the stub, which is to store the auth code's state in the database. We now generate the auth code, rather than using a hard-coded value. Note, however, that this is not wired together to be the auth code that we return on post - this will come in a subsequent commit + PR. The value of getting this in, for now, is to have a shared database definition to work from.