AUT-3862: Record `sub` from `/authorize` request
The `sub` field is sent to `/authorize` and we need to return it when the `/reverification` API is requested.

Here we pass it though the `/authorize` form as a hidden field and store it in the Reverification table as the return object.

This object will eventually be returned by the `/reverification` API.