BAU: Retrieve access token in `/reverification`
Using the access token passed in via the Authorisation header as a bearer token, look up the result from DynamoDB and return to the client.

Bearer token is expected as https://github.com/govuk-one-login/authentication-api/commit/161a57673d929a2348c464b884ccf6d5b707efe9 implements sending it in the `ReverificationResultsHandler`.

If the access token is missing, return an `invalid_token` error in accordance with RFC6750 "OAuth 2.0 Bearer Token Usage" Section 3.1: https://www.rfc-editor.org/rfc/rfc6750.txt (according to spec we should not present the reason if it's missing however this same reason will be used for when no result is recorded with the presented token as a shortcut for implementing a full OIDC server for the stub).