AUT-3982: Render full decrypted request.
Render the full decrypted request, including the access storage
token.  Rename the validator functions to better reflect what they do.

Tidy up by removing the ROOT_URI const which is no longer needed as
we retrieve it from the request.

Use pretty-json to render the full request.  Even though json-pretty
has a bug that causes the storage access token claim to be rendered in
correctly it still improves the presentation of the request and
makes it easier to read.