AUT-3822: Add config for build environment.
Add new configuration for the build environment.  Added the Version ID
of the new secrets created in build.  We are specifying version numbers
rather than using SecretString to retrieve the secret as creating a new
version of the secret will then mandate a change to the sam template
that will, in turn, cause the lambda that consumes the secret to be
re-deployed picking up the new version of the secret.