BAU: Allow autosubmitting the orch stub form
Mission here is to enable using a hosted orch stub when running the frontend locally. We currently can start a journey, but a redirect back to the stub fails as cookies are not set correctly. We will update our orch stub proxy to redirect the user to the hosted orch stub which will set cookies correctly, to instead of directly POSTing to this stub.

This part allows the form to be auto submitted on page load based on a query string value. This will be used by the orch stub proxy in the frontend to prevent an unnecessary extra form submission.